add upgrade parameters

+++ Assets/Scripts/Data/ObjectWithProductLevel.cs
@@ -5,7 +5,7 @@ public class ObjectWithProductLevel
 {
     public int levelUpPrice;
     public int maxProductAmount;
-    public float productPrice;
-    public int durationOfSelection;
+    public int productPrice;
+    public float durationOfSelection;
 }
